+++ src/hades.rs
@@ -26,6 +26,7 @@ mod strategies;
 
 use mds_matrix::MDS_MATRIX;
 use round_constants::ROUND_CONSTANTS;
+use strategies::Strategy;
 
 const TOTAL_FULL_ROUNDS: usize = 8;
 
@@ -36,9 +37,9 @@ const CONSTANTS: usize = 960;
 /// The amount of field elements that fit into the hades permutation container
 pub const WIDTH: usize = 5;
 
+pub(crate) use strategies::permute;
 #[cfg(feature = "zk")]
-pub use strategies::GadgetStrategy;
-pub use strategies::{ScalarStrategy, Strategy};
+pub(crate) use strategies::permute_gadget;
 
 const fn u64_from_buffer<const N: usize>(buf: &[u8; N], i: usize) -> u64 {
     u64::from_le_bytes([