+++ src/perm_uses.rs
@@ -8,7 +8,7 @@
 
 use dusk_bls12_381::BlsScalar;
 
-use crate::hades::{ScalarStrategy, Strategy, WIDTH};
+use crate::hades::{permute, WIDTH};
 
 /// Takes in one BlsScalar and outputs 2.
 /// This function is fixed.
@@ -23,7 +23,7 @@ pub fn two_outputs(message: BlsScalar) -> [BlsScalar; 2] {
     // Since we do a fixed_length hash, `words` is always the size of `WIDTH`.
     // Therefore, we can simply do the permutation and return the desired
     // results.
-    ScalarStrategy::new().perm(&mut words);
+    permute(&mut words);
 
     [words[1], words[2]]
 }